docker/_archive/trader/requirements.txt
websockets
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.1
tastytrade-sdk

